重构：重新组织 gem 文件.
