+++ .ruby-version
@@ -1,0 +1,1 @@
+ruby-2.3.1
