+++ examples/demo.rb
