+++ lib/sensors_analytics/version.rb
@@ -1,0 +1,4 @@
+
+module SensorsAnalytics
+  VERSION = '1.5.3'
+end
