+++ lib/sensors_analytics_sdk.rb
@@ -1,0 +1,9 @@
+
+require 'sensors_analytics/version'
+require 'sensors_analytics/errors'
+require 'sensors_analytics/consumers'
+require 'sensors_analytics/client'
+
+module SensorsAnalytics
+
+end
