--- sensors_analytics_sdk/sensors_analytics_sdk.gemspec
@@ -1,11 +1,0 @@
-Gem::Specification.new do |s|  
-  s.name        = 'sensors_analytics_sdk'  
-  s.version     = '1.5.3'  
-  s.date        = '2016-10-14'  
-  s.summary     = "SensorsAnalyticsSDK"  
-  s.description = "This is the official Ruby SDK for Sensors Analytics."  
-  s.authors     = ["Yuhan ZOU"]  
-  s.email       = 'zouyuhan@sensorsdata.cn'  
-  s.files       = ["lib/sensors_analytics_sdk.rb"]  
-  s.homepage    =  'http://www.sensorsdata.cn'  
-end
